improve quantile nulls
